Creates a bespoke SQLite database for the finRes universe

Creates a SQLite database bespoke to the finRes universe. In particular, the database is designed to store Bloomberg data retireved using the pullit package.

db_create(path = NULL, n = 10L, verbose = TRUE)

Arguments

path
a scalar chatacter vector. Specifies target directory for the database file. Defaults to home directory.

n
a scalar integer integer vector. Specifies the number of database tables for create for historical data storage. This is a storage versus time trade-off paramater. More tables occupy more space on disk but offer better read/write performance. Defaults to 10.

verbose
a logical scalar vector. Should progression messages be printed? Defaults to TRUE.

Details

Creates a SQLite database bespoke for use within the finRes suite context. The database is created in the destination directory specified in path. Populate with db_store.

See also

Examples

# NOT RUN {
db_create()
# }