Snapshots the content of a storethat SQLite database.

Provides a snapshot of the data stored in an existing storethat SQLite database.

db_snapshot(file = NULL, instrument, book = "all", name = "all")

Arguments

file
a scalar character vector. Specifies the path to the appropriate 'storethat.sqlite' file.

instrument
a scalar character vector. Specifies the financial instruments to get a snapshot for. Must be one of 'all', equity', 'index', 'fund' or 'futures'.

book
a scalar character vector. Instrument dependent; for a given instrument, specifies the book for the snapshot; 'all' snapshots all the books available for the given instrument.

name
a scalar character vector. Instrument dependent; for a given instrument, specifies the name for the snapshot; 'all' snapshots all the names available for the given instrument.

See also

Examples

if (FALSE) db_snapshot()